Fixed interface unzipping bug, vertex enrichment strategy added
Issue in interface unzip was not all elements with interface nodes
necessarily have interface side. Neighbors of elements with interface
sides were also unzipped.

For each vertex, a Vertex enrichment data structure is setup which
allows for the node to know which basis interpolates into it, the
enrichment level of that basis, and the weights.