Parallel Ghost
Majority of this commit supports communicating T-matrices across
processors  for interpolation vertices in the aura that are needed for
the ghost. This requires adding basis functions that do no exist on the
partition to the partition that interpolate into this aura vertex. Thes
vertices do not have a background vertex interpolation and will error if
this information is needed. Note: not all use cases of ghost in parallel
are sufficiently robust and I have experienced some examples failing. It
does seem to work for the fluid problem with spheres.

Added the first id to max+1
Added a global to local enriched basis map in the interpolation mesh