fixed bug due to which exo writer crashed in XTK
Side sets can snap to to background facets on the domain boundary
these facets are part of the aura and therefore the XTK enr. IG mesh
but not the exo mesh if the aura element is the leader. added checks
that exclude these elements from being skipped consistently in the
exo writer