[IWG] Utilize nonlinear variants for bulk and dirichlet IWG.
For the bulk and dirichlet IWGs:
The implementation for the linear and nonlinear variants are identical.
For the bulk a minor modification was necessary,
see previous commit: 78b351aa4680ca1a6f3d026a8e49d1c0f9eda41d

Therefore, we can call the nonlinear variants and get rid of the linear
ones.