[Contact] Adds the mFollowerPreviousFIManager to the FEM set class.
For contact problems we want to calculate the relative movement between
two contact surfaces. To do so, we need to evaluate the displacement
fields on the leader and follower side at the current and the previous
time step.
To prepare this, the Follower Field Interpolator at the previous time
step is added the FEM set class.
Additionally, a call to evaluate the previous time step pdof values is
added when accessing the nodal displacements in the FEM set class.