[Contact] Adds FI manager setup for prev. timestep in Interpol El.
In the Interpolation Element class, the field interpolation
coefficients are setup for current and previous Field Interpolators.
The previous FI are only setup in the case of time continuity.
The previous leader FI was already setup (as it is used in
time-continuity).
The setup of the previous folloer FI is added here.

This was missed in commit: 58b2f8c