[Contact/FI] Fixes storage of follower side for time continuity.
If there is time continuity we store the current values, such that they
are accessible as "previous" in the next step. The follower side was
storing the leader values, as the follower specification was missing.
In combination with the previous commits, a double-sided side-set can
now access its previous time-step follower values.