[Contact] Introduces relative velocity/slip to Mlika Contact IWG.
The relative tangential velocity (and slip increment) as defined in:
Poulios and Renard - 2015 - An unconstrained integral approximation
of large sliding frictional contact between deformable solids -
equation (25)
is added to the Mlika contact IWG.
This quantity is fundamental to evaluate tangential contact
contributions.

Note: The friction coefficient is set to zero, deactivting the added
functionality, as it is work in progress (and not to interfere with the
existing test).