[Contact] Updates residual for frictional case in Mlika contact IWG.
In Mlika et al. - 2017 - An unbiased Nitsche's formulation ...
the Nitsche's formulation for frictional contact is given in (15).
For the penalty case: \hat{\sigma}_N = 0 and \gamma = 0
the expression reduces to:
D J(u) + 0.5 \int C(..) * D( u(Y) - u(X) )
where C(..) is the friction law and D( u(Y) - u(X) )
is the linearization of the gap vector: tdGapvecdu and tdGapvecdv

Note: the friction coefficient is still set to zero, deactivating this
branch (as it is under development).