[Contact] Updates the Nitsche contact unit test to frictional mechanics.
Adds "previous time step" objects and fields to the contact unit test,
such that the contact IWG is able to evaluate frictional contributions
(which utilize fields from the previous time-step).

Note: whether friction is used or not is currently hard-coded in the
Mlika contact IWG. Needs to be added as a parameter to the IWG first,
then added here as well.