[Contact] Fixes bug in outward normal for linear deformed geometry.
Bug was introduced in commit: 99a5ff2c1ec6340a8363dafbdb09a707a16bab24
In that commit the evaluation of the normal for linear deformed
and consistent case was moved to a subroutine.
The bug was only introduced in the linear deformedcase.

The line previously read:
Matrix< DDRMat > tLeaderdNormaldU( tSpaceDim, tNumDofs );
i.e. variable declaration with size.

It was changed to: tLeaderdNormaldU( tSpaceDim, tNumDofs );
i.e. access to an element that is outside the bounds.

The setup of the variables is now done outside the subroutines.

Fixes failing unit test for contact with linear deformed geometry:
IWG_Struc_Nonlinear_Contact_Nitsche_NEOHOOK_MLIKA_SYM_LinDisp