[Contact] Fixes remapping unit test.
In the unit test the function abs(..) was used to get the absolute value.
Unfortunately, without the std:: (as in std::abs(..) ), the C-style
variant of abs is used, which is designed for integers.
The easy fix is to add std:: , this way the C++ variant is used, which
is overloaded with various data types.