[Contact] On the implementation of the residual for pure penalty.
In case of a pure penalty formulation, the scaling of all terms related
to traction (mTractionScaling) and the parameter mTheta are both zero.
In comparison to the full Nitsche formulation, the penalty regularized
case is significantly simpler.
Therefore, to derive the jacabion, the penalty case is chosen as
starting point.
To start off, all terms related to normal (frictionless) contact are
removed (commented out) from the jacobian.
In the follow up commits, the residual for the frictional contribution
is added step-by-step, as is the related jacobian.